requirement.txt
requests
beautifulsoup4

elasticsearch

pydantic

google-cloud-aiplatform
google-generativeai
